pubmed-agent/requirements.txt
requests>=2.31.0
openai>=1.12.0
PyYAML>=6.0.1
APScheduler>=3.10.4
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
